[sledge] Function summarization: solver can show pre
Summary:
This diff is preparation for function summarization and focuses on
function calls and function summary precondition computation.

It introduces `-function-summaries` flag behind most of functionality is
hidden, when enabled on each call

* A function summary is computed by quantifying all the non-formal/global variables
  and removing all the segments that are not reachable from them
* `pre` and `foot` are computed from function summary and the calling context
  by replacing formals with actuals again.
* A solver is asked if `pre` entails `foot` and a frame is printed if it
  does

Currently this only works for formulas without disjunctions, so when
function summaries are enabled, that state is first moved to dnf and then
the call is done for each disjunct.

Reviewed By: ngorogiannis

Differential Revision: D15898928

fbshipit-source-id: 49d32504c