Remove analyze_ondemand from Ondemand.callbacks
Summary:
Currently, `Callbacks.analyze_procedures` creates a function to call the method `Callbacks.iterate_procedure_callbacks`. This is supplied as an argument to functions in `ondemand.ml`, so that it can be invoked. This is done to avoid a cyclic dependancy.

This diff moves the functions that `ondemand.ml` needs to call into `ondemand.ml`, avoiding the need to supply them as arguments.

Reviewed By: ngorogiannis

Differential Revision: D16028836

fbshipit-source-id: 16ae27a3e