[pulse] allow models in invalidation traces
Summary:
Be more flexible in what type of function calls are allowed in `ViaCall ...` actions to be able to include models.

Also get rid of `here here` in traces /o\

As a side-effect, get more precise (=qualified) procedure names in
traces (but not in messages so as not to be too verbose).

Reviewed By: mbouaziz, ngorogiannis

Differential Revision: D16121092

fbshipit-source-id: fb51b02f8