[changed files] don't fail on non-existent absolute paths
Summary:
When a file is passed to infer through `--changed-files-index` which is
- an absolute path
- the file does not exist

Then the code fails throwing an exception in the function below while trying to relativise the absolute path.

The behaviour on relative paths is to skip missing files, and does not fail because Infer does not attempt to relativise them.

Swallow the exception and skip the file; and so unify the behaviour across relative and absolute paths.

Reviewed By: mityal

Differential Revision: D16279672

fbshipit-source-id: 33b468da7