[cost] Revise hasNext() to avoid bottom in condition
Summary:
Before this diff it returned `[0,size-1]`, but which became bottom
when size was given by 0.  As a result, it made the both branches of
`if(iterator.hasNext())` unreachable.  Similarly, if the size was 1,
it only visited the false branch of the if condition because the
condition value was `[0,0]` at that time.

This diff changes it to return `[0,size]`, so that

* the false branch is reachable when the size is 0
* the both branches are reachable when the size is 1

Reviewed By: ezgicicek

Differential Revision: D16803000

fbshipit-source-id: f8772be27