[annot reachability] apply sanitizers in more cases
Summary:
Change the logic of the annotation reachability checker in the following
ways:

1. Sanitizers take priority over sinks, i.e. a procedure that is both a
sink and a sanitizer is not a sink. This changes the existing tests
that seemed to assume the opposite. However I think that way is more
useful and goes better with the fact that sanitizers are specified as
"overrides".

2. When applying a summary, check again that we are not in a sanitizer
for the corresponding sink.

Without (2) this there was a subtle bug when several rules were
specified.  For example, if `sink_wrapper()` wraps `sink()` for a rule
`R` then the summary of `sink_wrapper()` will be: `R-sink : call to sink()`.

Then, suppose `sanitizer()` calls `sink_wrapper()` and `sanitizer()` is
a sanitizer for `R` but not for another rule `R'`. The previous code
would add the call to `sink()` to the summary of `sanitizer()` because
it's not a sanitizer for `R'`, even though `sink()` is not a sink for
`R'`!

The current code will re-apply the rules correctly so that sinks are
matched only against the right sanitizers.

Reviewed By: skcho

Differential Revision: D16895577

fbshipit-source-id: 266cc4940