[racerd] output access expressions language-sensitively
Summary:
Use whatever information we can to decide whether to use C or Java
syntax when outputting an access expression, now that we store them as
such.

Also, make cluster callbacks explicitly set the language, as this was not done before and led to some confusion (Clang being set when analysing a Java file).

Reviewed By: skcho

Differential Revision: D16884160

fbshipit-source-id: 40adf9f35