[clang] wrap test determinator in frontend exception catcher
Summary:
The clang frontend has bugs. When a bug we know about happens some
exception is raised and, most of the time, logged away so as not to
crash the whole process. This catching of exceptions wasn't done from
testDeterminator so it could crash where capture didn't. This diff wraps
the crashy function in test determinator to avoid that.

Reviewed By: ngorogiannis

Differential Revision: D16963178

fbshipit-source-id: 87a4ff70b