Clean specs directory before running capture
Summary:
Currently, the specs directory is cleaned after running capture. This means that the `changed-files` are interpreted in the context of the second set of source files. Therefore if a procedure is deleted from the second set of source files, its specs file will not be deleted.

This moves the cleaning of the specs directory to before capture, to avoid this problem.

Reviewed By: ngorogiannis

Differential Revision: D17093980

fbshipit-source-id: e1a8d8a54