[nullsafe] remove 'annotation' abstraction.
Summary:
This abstraction was not always used consistently.

Its usage made more sense when it supported both present annotations and
optional annotation (which got removed in previous diff).

The rought semantic of that was "what is the inferred type for such and
such value (variable or expression) in typestate". So it is not really
_annotation_ in first place, it is more like "what we inferred about
nullability given annotations, known special cases, and current sybmolic
execition state".

Let's explicitly rename `map` to `is_nullable`. If/when we need to
enhance this further (and we likely will), we will do it accordingly.

Reviewed By: jvillard

Differential Revision: D17153434

fbshipit-source-id: 3c85b56df