[nullsafe][refactor tests] split and improve FieldNotNullable tests
Summary:
See motivation below.
This diff is dealing with FieldNotNullable:
- move not relevant subclasses into dedicated classes and files
- modify the tests so they comply with the standards below

--Motivation--

Gradual mode we are going to introduce is an invasive change in how Infer
treats nullability semantics.

In order to make the change in a controllable way, we need the tests to comply with the
following standards and conventions.

1. For each code peace where we expect a bug to happen, the there should be
corresponding (minimally different from above) peace of code where we expect a bug to NOT happen. (This is to ensure bug is happening for exact reason we think it is happening).
2. Conversely: for each peace of code where we expect a bug to be NOT
present, there shuold be a peace of code where the bug IS happening.
(Otherwise there can be too many reasons for a bug NOT to happen).
3. Convention: end corresponding methods IsOK and IsBUG correspondingly.
4. Keep code examples as small as possible.

Reviewed By: ngorogiannis

Differential Revision: D17183222

fbshipit-source-id: 83d03e67f